Linear Programs Task2

+++ src/LinearPrograms/Task1.java
@@ -1,8 +1,8 @@
 package LinearPrograms;
 // 1. Найдите  значение функции: z = ( (a – 3 ) * b / 2) + c.
 public class Task1 {
-    public static double task1 (int a, int b, int c) {
-        double z = (a - 3) * (double) b / 2 + c;
+    public static double task1 (double a, double b, double c) {
+        double z = (a - 3) * b / 2 + c;
         return z;
     }
 }
